+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Backward-incompatible changes:
 Changes:
 ^^^^^^^^
 
+- UNIX epoch timestamps from ``structlog.processors.TimeStamper`` are more precise now.
 - Add *repr_native_str* to ``structlog.processors.KeyValueRenderer`` and ``structlog.dev.ConsoleRenderer``.
   This allows for human-readable non-ASCII output on Python 2 (``repr()`` on Python 2 haves like ``ascii()`` on Python 3 in that regard).
   As per compatibility policy, it's on (original behavior) in ``KeyValueRenderer`` and off (humand-friendly behavior) in ``ConsoleRenderer``.